+++ client/modules/Service/MDServiceSelector.jsx
@@ -7,50 +7,72 @@ MDServiceSelector = React.createClass({
 	mixins:[ReactMeteorData],
 
 	getMeteorData() {
-		var request,service,subservice,services,subservices;
-		request = this.props.context;
-		if(request&&request.facility) {
-			services = request.facility.servicesRequired;
-		}
-		service = this.props.value||{};
-		subservice = service.subservice;
-		if(service.children) {
+
+		let request = this.props.context,
+			facility = null,
+			service = this.props.value,
+			subservice = null,
+			services = null,
+			subservices = null;
+
+		if(request) {
+			if( request.facility && request.facility._id ) {
+				facility = Facilities.findOne(request.facility._id);
+				services = facility.servicesRequired;
+			}
+			if(request.subservice) {
+				subservice = request.subservice;
+			}
+		} 
+
+		if(service && service.children) {
 			subservices = service.children;
 		}
 
 		return {
-			services:services,
-			subservices:subservices,
 			service:service,
+			services:services,
 			subservice:subservice,
+			subservices:subservices,
 		}
 	},
 
-	updateSubService(newVal) {
-		var service = this.props.value||{};
-		service.subservice = newVal;
-		this.props.onChange(service);
+	updateService(newService) {
+		let update = {
+			service: newService,
+			subservice: null
+		}
+		if(newService.data&&newService.data.supplier) {
+			update.supplier = Teams.findOne(newService.data.supplier._id);
+		}
+		if(this.props.multiChange) {
+			this.props.multiChange(update);
+		}
+	},
+
+	updateSubService(newSubService) {
+		let update = {
+			subservice: newSubService
+		}
+		if(newSubService.data&&newSubService.data.supplier) {
+			update.supplier = Teams.findOne(newSubService.data.supplier._id);
+		}
+		if(this.props.multiChange) {
+			this.props.multiChange(update);
+		}
 	},
 
 	render() {
 		var shouldIncludeFacilitySelector = this.props.options.shouldIncludeFacilitySelector;
 		return (
 			<div className="row">
-				{
-					shouldIncludeFacilitySelector?
-						<div className="col-md-12">
-							<MDFacilitySelector onChange={this.facilitySelected}/>
-						</div>
-					:
-						null
-				}
 				<div className="col-md-6">
 					<AutoInput.MDSelect 
 						items={this.data.services} 
 						disabled={!this.data.services}
 						selectedItem={this.data.service}
 						itemView={NameCard}
-						onChange={this.props.onChange}
+						onChange={this.updateService}
 						placeholder="Service"
 					/>
 				</div>