+++ client/modules/Service/ServiceViewDetail.jsx
@@ -15,50 +15,7 @@ ServiceViewDetail = React.createClass({
         return data;
     },
 
-    createRule(newRule) {
-        console.log(newRule);
-        var facility = newRule.facility;
-        if(facility) {
-            var services = facility.servicesRequired;
-            //get index of the selected service
-            var idx=-1;
-            for(var i in services) {
-                if(newRule.service&&newRule.service.name&&services[i].name==newRule.service.name) {
-                    idx = i;
-                    break;
-                }
-            }
-            if(idx>=0) {
-                var service = services[idx];
-                console.log({service,idx});
-                if(!service.data) {
-                    service.data = {};
-                }
-                if(!service.data.complianceRules) {
-                    service.data.complianceRules = [];
-                }
-                //to avoid circular search remove facility and then add with just name and _id
-                var copy = _.omit(newRule,'facility','service','event');
-                if(newRule.facility) {
-                    copy.facility = _.pick(newRule.facility,'name','_id');
-                }
-                if(newRule.event) {
-                    copy.event = _.pick(newRule.event,'name','_id');
-                }
-                if(newRule.service) {
-                    copy.service = _.pick(newRule.service,'name');
-                }
-                //console.log(copy);
-                service.data.complianceRules.push(copy);
-                services[idx] = service;
-            }
-            facility.setServicesRequired(services);
-        }
-        Modal.hide();
-    },
-
     deleteRules(serviceName) {
-        console.log(serviceName);
         var services = this.data.facility.servicesRequired;
         var idx=-1;
         for(var i in services) {
@@ -71,19 +28,6 @@ ServiceViewDetail = React.createClass({
             services[idx].data.complianceRules = null;
             this.data.facility.setServicesRequired(services);
         }
-    },
-
-    handleCreateRuleClick() {
-        Modal.show({
-            content:<AutoForm 
-                item={{
-                    facility:this.data.facility,
-                    service:this.props.item
-                }}
-                schema={ComplianceRuleSchema} 
-                onSubmit={this.createRule}
-            />
-        })
     },
 
     render() {
@@ -117,7 +61,7 @@ ServiceViewDetail = React.createClass({
                             <span>{results.servicesFailed} non-compliant services</span>
                         </div>
                         <div style={{textAlign:"right"}}>
-                            <button onClick={this.handleCreateRuleClick} className="btn btn-flat" style={{backgroundColor:"transparent",color:"#fff",padding:"10px 20px 0px 20px"}}>New Rule</button>
+                            <button onClick={()=>{FABActions.createComplianceRule()}} className="btn btn-flat" style={{backgroundColor:"transparent",color:"#fff",padding:"10px 20px 0px 20px"}}>New Rule</button>
                         </div>
                     </div>                    
                 </div>