+++ lib/startup.jsx
@@ -1,4 +1,5 @@
 if(Meteor.isClient) {
 	Meteor.subscribe('teamsAndFacilitiesForUser');
 	Meteor.subscribe('users');
+	Meteor.subscribe("contractors");
 }